requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.18.1
annotated-types==0.7.0
anyio==4.12.1
apache-airflow==3.1.6
apache-airflow-core==3.1.6
apache-airflow-providers-common-compat==1.11.0
apache-airflow-providers-common-io==1.7.0
apache-airflow-providers-common-sql==1.30.2
apache-airflow-providers-smtp==2.4.1
apache-airflow-providers-standard==1.10.2
apache-airflow-task-sdk==1.1.6
argcomplete==3.6.3
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
backports.strenum==1.3.1
blinker==1.9.0
cachetools==6.2.4
cadwyn==5.4.6
category_encoders==2.7.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.4
dash==3.3.0
databricks-cli==0.18.0
databricks-sdk==0.78.0
decorator==5.2.1
Deprecated==1.3.1
deprecation==2.1.0
dill==0.4.0
dnspython==2.8.0
docker==6.1.3
email-validator==2.3.0
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.117.1
fastapi-cli==0.0.20
fastjsonschema==2.21.2
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenback==1.3.0
greenlet==3.3.0
grpcio==1.76.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==7.2.1
iniconfig==2.3.0
ipython==8.38.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kaleido==1.2.0
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
libcst==1.8.6
lightgbm==4.6.0
linkify-it-py==2.0.3
llvmlite==0.46.0
lockfile==0.12.2
logistro==2.0.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.5
matplotlib-inline==0.2.1
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.9.2
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
more-itertools==10.8.0
msgspec==0.20.0
narwhals==2.15.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.63.1
numpy==1.26.4
oauthlib==3.3.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.5
pathspec==1.0.3
patsy==1.0.2
pendulum==3.1.0
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
plotly-resampler==0.11.0
pluggy==1.6.0
pmdarima==2.0.4
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycaret==3.3.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyod==2.0.6
pyparsing==3.3.1
pytest==9.0.2
pytest-timeout==2.4.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
python-slugify==8.0.4
pytz==2023.4
PyYAML==6.0.3
querystring-parser==1.2.4
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.17.1
rpds-py==0.30.0
rsa==4.9.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
setproctitle==1.3.7
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sklearn-compat==0.1.5
sktime==0.26.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.6
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tbats==1.1.3
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.4.0
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
widgetsnbextension==4.0.15
wirerope==1.0.0
wrapt==2.0.1
wurlitzer==3.1.1
xxhash==3.6.0
yellowbrick==1.5
zipp==3.23.0
